[VM] merge etienne work

+++ include/vm.h
@@ -6,7 +6,7 @@
 /*   By: amartinod <amartino@student.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/18 12:52:38 by amartinod         #+#    #+#             */
-/*   Updated: 2020/07/17 14:12:21 by amartinod        ###   ########.fr       */
+/*   Updated: 2020/07/18 15:36:00 by amartinod        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -113,7 +113,7 @@ uint8_t		get_param(t_vm *vm, t_cursor *cursor, size_t pc);
 uint8_t		param_type(t_vm *vm, t_cursor *cursor, uint8_t bit_shift);
 int32_t		get_register(t_cursor *cursor, size_t reg_number);
 int32_t		get_indirect(t_vm *vm, t_cursor *cursor, uint8_t nb_arg);
-int32_t		modulo(int32_t a, int32_t b);
+uint16_t	modulo(int16_t a, int16_t b);
 int32_t		get_param_when_3_possible_type(t_vm *vm, t_cursor *cursor, 
 									uint8_t arg, uint8_t type);
